Update RecordArray to work with Ember.Sortable
Ember.Sortable uses array.slice, which uses the current length as the
maximum number of elements that can be sliced.  Length is based on the
length of `storeKeys`.  This commit fixes the notifications to send
arrayWillChange before storeKeys changes and arrayDidChange afterwards.
This fixes array.slice, and thus fixes sortable array proxies.